+++ src/Markdig.SyntaxHighlighting.Portable/LanguageTypeAdapter.cs
