+++ src/Markdig.SyntaxHighlighting.Portable/Markdig.SyntaxHighlighting.nuspec
