+++ src/Markdig.SyntaxHighlighting.Portable/Properties/AssemblyInfo.cs
