+++ src/Markdig.SyntaxHighlighting.Portable/SyntaxHighlightingCodeBlockRenderer.cs
