+++ src/Markdig.SyntaxHighlighting.Portable/SyntaxHighlightingExtension.cs
