+++ src/Markdig.SyntaxHighlighting.Portable/SyntaxHighlightingExtensions.cs
