+++ src/Markdig.SyntaxHighlighting.Tests.old/LanguageTypeAdapterTests.cs
