+++ src/Markdig.SyntaxHighlighting.Tests.old/Properties/AssemblyInfo.cs
