+++ src/Markdig.SyntaxHighlighting.Tests.old/SyntaxHighlightingExtensionsTests.cs
